Send a key_exchange_failed message
It would be nice if we could just drop messages if the public key hash
verification fails, but this means either the public key was tampered
with in flight or the user entered their password incorrectly.

We need to send this message so the UI on the website can reset and
prompt the user for their password again.