Prevent NPE when starting DeviceFinderService
I've been getting some occasional NPEs here.  This always happens when
you start a new location request while the DeviceFinderService is
already running.  I'll have to look into why the Intent is null, I have
a hunch that it has to do with the fact that the session is started as
STICKY.