Save account password if account exists
When the user changes their password and we prompt them to login again,
we end up following the same codepath as a new account.  This causes the
password to not be updated in the account manager.  We can check to see
if the account already existed when adding it.  If it exists, just
update the password.