Check minimum required application version
When the application first boots, makes a request to determine the
minimum required application version for compatibility with the server.
If the version is not compatible, fires a notification which shows a
message telling the user to update to the latest version of CyanogenMod.

When logging in, after WiFi is enabled, also check the minimum version.

In most cases, the version check will be completed before login, if the
device is connected to a network.

Change-Id: Ic0677088366453276fa48b7d82d697ceecdd9416